chore: add script to automatically update plunker bundle
